chore(extract-utils): Extract utility functions from ME into separate module.
