+++ mokhaUtils.py
@@ -1,0 +1,26 @@
+def printOptions(options, key=None):
+    for i, option in enumerate(options):
+        if(key == None):
+            print("%s: %s" % (i + 1, option))
+        elif(key in option.keys()):
+            print("%s: %s" % (i + 1, option[key]))
+        else:
+            print("Invalid Key")
+
+
+def getSelection(options, key=None):
+    printOptions(options, key)
+    while True:
+        try:
+            selection = int(
+                input("Select an option: "))
+            if(selection <= len(options)):
+                break
+            else:
+                print("ERROR: Please enter a value inside the array.")
+        except ValueError:
+            print("Please input an integer value.")
+    try:
+        return options[selection - 1]
+    except:
+        print("I got an error... I sense a disturbance in the force.")
